Make Modifiers untyped only.
The typed variant is no longer needed. This means modifiers can safely be
ignored in typed trees if we so choose.
